Renamed the suppliers app to companies
Renamed suppliers to companies, as it makes sense to be companies rather than suppliers who are persons.
