+++ spare-pal/apps/companies/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+# Register your models here.
+
+from apps.companies.models import Company
+
+admin.site.register(Company)
