+++ spare-pal/apps/companies/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class SuppliersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'apps.suppliers'
+    name = 'apps.companies'
