--- spare-pal/apps/suppliers/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-
-from apps.suppliers.models import Supplier
-
-admin.site.register(Supplier)
